uninit/init split all the things
