+++ facet-reflect/src/poke/iset.rs
@@ -1,0 +1,47 @@
+/// Keeps track of which fields were initialized, up to 64 fields
+#[derive(Clone, Copy, Default)]
+pub struct ISet(u64);
+
+impl ISet {
+    /// Creates a new ISet with all (given) fields set.
+    pub fn all(fields: &[Field]) -> Self {
+        let mut iset = ISet::default();
+        for (i, _field) in fields.iter().enumerate() {
+            iset.set(i);
+        }
+        iset
+    }
+
+    /// Sets the bit at the given index.
+    pub fn set(&mut self, index: usize) {
+        if index >= 64 {
+            panic!("ISet can only track up to 64 fields. Index {index} is out of bounds.");
+        }
+        self.0 |= 1 << index;
+    }
+
+    /// Unsets the bit at the given index.
+    pub fn unset(&mut self, index: usize) {
+        if index >= 64 {
+            panic!("ISet can only track up to 64 fields. Index {index} is out of bounds.");
+        }
+        self.0 &= !(1 << index);
+    }
+
+    /// Checks if the bit at the given index is set.
+    pub fn has(&self, index: usize) -> bool {
+        if index >= 64 {
+            panic!("ISet can only track up to 64 fields. Index {index} is out of bounds.");
+        }
+        (self.0 & (1 << index)) != 0
+    }
+
+    /// Checks if all bits up to the given count are set.
+    pub fn are_all_set(&self, count: usize) -> bool {
+        if count > 64 {
+            panic!("ISet can only track up to 64 fields. Count {count} is out of bounds.");
+        }
+        let mask = (1 << count) - 1;
+        self.0 & mask == mask
+    }
+}
